shopman-core/stocking/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "shopman-stocking"
version = "0.3.0"
description = "Shopman Stocking — Inventory Management"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.2,<6.0",
    "shopman-utils",
]

[project.optional-dependencies]
api = [
    "djangorestframework>=3.14",
]
unfold = [
    "django-unfold>=0.80,<1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.5",
    "djangorestframework>=3.14",
]

[tool.setuptools.packages.find]
include = ["shopman.stocking*"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "stocking_test_settings"
pythonpath = ["."]
